fix: fix -Wextra warnings across kernel_types, ttnn/ (r23 batch)
- kernel_types.hpp: add = {} defaults to DataMovementConfig and ComputeConfig
  container fields (compile_args, defines, named_compile_args, etc.)
- program_descriptors.hpp: add = {} defaults to CBDescriptor format_descriptors
- matmul_program_config_types.hpp: add = std::nullopt to fused_activation fields
- tensor_impl.cpp: fix int64_t vs size_t in strides.size() comparison
- tensor/xtensor/partition.cpp: cast dim to size_t vs tensor_dims
- distributed_tensor.cpp: cast cluster_axis.value() to size_t
- reports.cpp: use uint32_t for page_index loop
- compute_throttle_utils.cpp: cast num_cores to uint32_t vs constexpr uint32_t
- reshape_common.cpp: use size_t for index loop
- op_profiler_json.cpp: cast json .value() to size_t vs packed_size

+++ ttnn/cpp/tools/profiler/op_profiler_json.cpp
@@ -99,7 +99,7 @@ static json get_kernels_json(ChipId device_id, const Program& program) {
         for (const auto& binary_meta : kernel.binary_meta) {
             auto key = fmt::format(
                 "{}_{}_{}_max_kernel_size", core_type_name, processor_class_name, binary_meta.processor_type);
-            if (kernelSizes.value(key, 0) < binary_meta.packed_size) {
+            if (static_cast<std::size_t>(kernelSizes.value(key, 0)) < binary_meta.packed_size) {
                 kernelSizes[key] = binary_meta.packed_size;
             }
         }
